% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{absolute_loss}
\alias{absolute_loss}
\title{absolute value of difference between model output and ground truth}
\usage{
absolute_loss(sim_output, ground_truth)
}
\arguments{
\item{sim_output}{vector of model outputs}

\item{ground_truth}{vector of data to compare to model output}
}
\description{
absolute value of difference between model output and ground truth
}
