% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_metaRVM}
\alias{calibrate_metaRVM}
\title{calibrate_metaRVM}
\usage{
calibrate_metaRVM(
  ground_truth,
  base_config,
  periods,
  parameter_bounds,
  npop = 6,
  param_offset = 0,
  n_checkpoints = 10,
  n_trajectories = 50,
  init_nsim = 500,
  grid_size = 500,
  sim_batch_size = 500,
  tot_nsim = 1500,
  checkpoint_dir = "./checkpoints",
  n_forecast_samples = 100,
  forecast_horizon = 4,
  moving_window_size = 4,
  calibration_strategy = c("moving_window", "distinct"),
  ncores = 2,
  verbose = T
)
}
\arguments{
\item{ground_truth}{output data to calibrate to}

\item{base_config}{configuration file containing the "static" (i.e., non-calibrated) parameters for \verb{\code{MetaRVM}}}

\item{periods}{named list where each entry is an integer vector of weeks to calibrate}

\item{parameter_bounds}{names list of parameter:c(minimum val, maximum val), for example list("ts"=c(0.01,0.5))}

\item{npop}{number of subpopulations (must be same size as mixing matrices)}

\item{n_checkpoints}{number of checkpoints to save}

\item{n_trajectories}{number of trajectories to save}

\item{init_nsim}{size of Latin Hypercube sample for initial simulations}

\item{grid_size}{size of Latin Hypercube grid during Thompson sampling}

\item{sim_batch_size}{number of simulations to acquire in each round of Thompson sampling}

\item{tot_nsim}{total number of simulations to run for calibration}

\item{checkpoint_dir}{directory to save checkpoint files}

\item{n_forecast_samples}{number of samples for forecast}

\item{forecast_horizon}{number of weeks to forecast for}

\item{moving_window_size}{when \code{calibration_strategy}="moving_window" size of moving window}

\item{calibration_strategy}{one of "moving_window" (default) or "distinct"}

\item{ncores}{number of cores to use for parallel runs (cannot be larger than parallel::detectCores() - 1)}

\item{verbose}{boolean option to print output during calibration}
}
\value{
a named list
\itemize{
\item \code{posterior}: \code{n_trajectories} of parameter values
\item \code{sims_post}: simulations from \code{posterior}
\item \code{checkpoints}: \code{data.table} with parameter id and stem of checkpoint file name
\item \code{sims_forecast}: simulations sampled from \code{sims_post} for \code{forecast_horizon} weeks
\item \code{config_files}: vector of configuration files with checkpoints for next period of calibration
}
}
\description{
calibrate metaRVM output to data
}
