% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{calibrate_week}
\alias{calibrate_week}
\title{calibrate_week}
\usage{
calibrate_week(
  np,
  npop,
  weeks,
  configs,
  ground_truth,
  parameter_bounds,
  lower_bound,
  upper_bound,
  init_nsim_per_config,
  sim_batch_size,
  grid_size = 100,
  tot_nsim = 100,
  error_function = NULL,
  verbose = F
)
}
\arguments{
\item{np}{number of dimensions to calibrate}

\item{npop}{number of populations (same dimension as mixing matrices)}

\item{weeks}{vector of weeks to calibrate to}

\item{configs}{vector of yaml file specifying potential configurations}

\item{ground_truth}{ground truth data to calibrate to}

\item{parameter_bounds}{named list of parameter -> range (example: ts: (0.01,0.50)) to be passed to run_sims}

\item{lower_bound}{vector of lower bounds for calibration parameters}

\item{upper_bound}{vector of upper bounds for calibration parameters}

\item{init_nsim_per_config}{number of simulations per config file}

\item{sim_batch_size}{number of sims to run per emulator acquisition}

\item{grid_size}{size of Latin Hypercube grid to explore per acquisition}

\item{tot_nsim}{total number of simulations to run}

\item{error_function}{(data,ground_truth) discrepancy function. Defaults to squared log loss.}

\item{verbose}{print output during calibration (default F)}
}
\value{
a named list with
\itemize{
\item \code{pars}: parameters acquired during calibration
\item \code{Y}: errors between simulations and data
\item \code{sims}: simulations corresponding to \code{pars}
}
}
\description{
Calibrate metaRVM model outputs to data. This function creates a Latin Hypercube sample (LHS) of the calibration parameters
and then adaptively explores parameter space using Thompson sampling. The algorith works as follows:
\itemize{
\enumerate{
\item Create an LHS of size \code{init_nsim_per_config}
\item Run simulations for those parameters
\item Assess the error between model outputs and data
\item Fit a Gaussian process emulator on the (parameters, error) observations
\item Perform Thompson sampling to acquire new parameters and run new simulations, refitting GPs each time
\item Return results of parameters, errors, and simulations
}
}
}
