% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateMetaRVM-package.R
\docType{package}
\name{calibrateMetaRVM-package}
\alias{calibrateMetaRVM}
\alias{calibrateMetaRVM-package}
\title{calibrateMetaRVM: Calibrating the MetaRVM Model}
\description{
A package to calibrate the MetaRVM model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RESUME-Epi/calibrate_MetaRVM}
  \item Report bugs at \url{https://github.com/RESUME-Epi/calibrate_MetaRVM/issues}
}

}
\author{
\strong{Maintainer}: Arindam Fadikar \email{afadikar@anl.gov} (\href{https://orcid.org/0000-0001-7396-0350}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item David O'Gara \email{dogara@anl.gov} (\href{https://orcid.org/0000-0002-1957-400X}{ORCID}) [contributor]
  \item Charles Macal [contributor]
  \item Martinez Moyano Ignacio Javier [contributor]
  \item Ozik Jonathan [contributor]
}

}
\keyword{internal}
