% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_surrogate.R
\name{draw_sample_composite_gp}
\alias{draw_sample_composite_gp}
\title{draw_sample_composite_gp}
\usage{
draw_sample_composite_gp(gp_list, newX, sim_batch_size)
}
\arguments{
\item{gp_list}{list of GPs}

\item{newX}{matrix of X locations}

\item{sim_batch_size}{size of sample}
}
\value{
matrix of Thompson sampling results
}
\description{
Draw sample from list of GPs
}
