% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_surrogate.R
\name{fit_composite_gp}
\alias{fit_composite_gp}
\title{Fit GP surrogate models for multiple configs}
\usage{
fit_composite_gp(X, Y, cat_inds, np)
}
\arguments{
\item{X}{A matrix of input parameters.}

\item{Y}{A vector of output values.}

\item{cat_ind}{A vector of categorical indicators.}
}
\value{
A list of fitted \code{laGP} objects, with names corresponding to the categories.
}
\description{
This function takes a matrix of inputs X, a vector of outputs Y, and a vector of categorical indicators cat_ind.
It divides Y and X according to cat_ind, and for each category, it fits a GP surrogate
using laGP. It then returns a list of the fitted laGP objects.
}
