% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_header_footer}
\alias{msg_header_footer}
\title{Messaging functions}
\usage{
msg_header_footer(n = getOption("width"))
}
\arguments{
\item{n}{size of window}
}
\description{
set of functions to control display output
}
