% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_surrogate.R
\name{predict_composite_gp}
\alias{predict_composite_gp}
\title{predict_composite_gp}
\usage{
predict_composite_gp(gp_list, newX)
}
\arguments{
\item{gp_list}{list of GPs from \code{fit_composite_gp}}

\item{newX}{matrix of new locations to predict on}
}
\value{
list of predictions for each GP
}
\description{
Make prediction for each GP in \code{gp_list} on \code{newX}
}
