% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{process_vac_data}
\alias{process_vac_data}
\title{Process vaccine data}
\usage{
process_vac_data(vac_dt, sim_start_date, sim_length, delta_t)
}
\arguments{
\item{vac_dt}{data.table of vaccines over time}

\item{sim_start_date}{starting date of simulation (such as 2025-10-01 for the first day of flu season)}

\item{sim_length}{number of timesteps to simulate}

\item{delta_t}{timestep size, defaults to 0.5 (half day)}
}
\description{
Process vaccine data
}
