% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{run_forecast_sims}
\alias{run_forecast_sims}
\title{run_forecast_sims}
\usage{
run_forecast_sims(
  pars,
  weeks,
  configs,
  metadata = NULL,
  return.full = F,
  week_offset = 0
)
}
\arguments{
\item{pars}{matrix of parameters}

\item{weeks}{vector of weeks to simulate for}

\item{configs}{vector of config files}

\item{week_offset}{increments the starting week (default 0)}

\item{return.F}{boolean to return full output (default F)}
}
\description{
run_forecast_sims
}
