% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_sim}
\alias{run_sim}
\title{MetaRVM Wrapper}
\usage{
run_sim(
  config,
  weeks,
  ts = NULL,
  pea = NULL,
  psr = NULL,
  do_chk = F,
  run_id = NULL,
  start_date = NULL,
  week_offset = 0,
  return.full = F
)
}
\arguments{
\item{config}{configuration file path or text}

\item{weeks}{Vector of weeks to simulate}

\item{ts}{Vector of transmissibility values (length = number of populations)}

\item{pea}{Vector of pre-existing immunity values (length = number of populations)}

\item{psr}{Vector of seasonal reduction values (length = number of populations)}

\item{do_chk}{checkpoint model}

\item{run_id}{}
}
\value{
Data table of weekly hospitalizations by age group
}
\description{
Wrapper function to run MetaRVM given parameters
}
