% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{run_sims}
\alias{run_sims}
\title{run batch of simulations}
\usage{
run_sims(
  pars,
  weeks,
  configs,
  metadata = NULL,
  return.full = F,
  week_offset = 0,
  parameter_bounds,
  npop,
  parallel = F,
  ncores = 2,
  debug = F
)
}
\arguments{
\item{pars}{matrix of calibration parameter values to simulate}

\item{weeks}{integer vector denoting weeks to simulate (example: weeks = 1:10)}

\item{configs}{vector of configuration files to run \code{MetaRBM}}

\item{metadata}{\code{data.table}}

\item{return.full}{boolean}

\item{week_offset}{integer}

\item{parameter_bounds}{named list}

\item{npop}{number of subpopulations}

\item{parallel}{boolean for parallel computing}

\item{ncores}{number of parallel cores to use}

\item{debug}{optional parameter used for debugging}
}
\value{
\code{data.table} of \code{MetaRVM} simulation results
}
\description{
run batch of simulations
}
