% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{squared_log_loss}
\alias{squared_log_loss}
\title{Loss functions
a collection of common loss functions}
\usage{
squared_log_loss(sim_output, ground_truth)
}
\arguments{
\item{sim_output}{simulation output}

\item{ground_truth}{ground_truth data}
}
\description{
sum of squared differences between model output and ground truth
}
