% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_surrogate.R
\name{update_composite_gp}
\alias{update_composite_gp}
\title{update_composite_gp}
\usage{
update_composite_gp(gp_list, newX, newY, new_cat_ind, m_sd_mat)
}
\arguments{
\item{gp_list}{list of GPs}

\item{newX}{matrix of new locations to fit}

\item{newY}{vector of new locations to fit}

\item{new_cat_ind}{vector of integers corresponding to configuration files}

\item{m_sd_mat}{matrix of mean, standard deviations to normalize newX and newY}

\item{updated}{\code{gp_list}}
}
\description{
Update each GP in \code{gp_list} with \code{newX} and \code{newY}
}
